+++ WebApplication/src/components/ResourceButtonGroup.vue
@@ -660,7 +660,7 @@ export default {
             }
         },
         hasApiViewButton() {
-            return  this.resourceLevel != 'bulk' && this.studiesSourceType == SourceType.LOCAL_ORTHANC
+            return  this.resourceLevel != 'bulk' && this.studiesSourceType == SourceType.LOCAL_ORTHANC && this.uiOptions.EnableApiView;
         },
         resourcesOrthancId() {
             if (this.resourceLevel == 'bulk') {