+++ release-notes.md
@@ -1,3 +1,9 @@
+Pending changes
+===============
+
+Fixes:
+  - Fixed the "UiOptions.EnableApiView" that was not taken into account.  The API View button was always visible.
+
 1.6.0 (2024-08-02)
 ==================